Allow Xcode 12 to make Catalina changes
and silence some warnings

+++ ApplicationController.m
@@ -52,7 +52,7 @@
 	[NSApp setServicesProvider:services];
 
 	// Force update the services menu if we have a new services version
-	int serviceVersion = [[NSUserDefaults standardUserDefaults] integerForKey:@"Services Version"];
+	NSInteger serviceVersion = [[NSUserDefaults standardUserDefaults] integerForKey:@"Services Version"];
 	if (serviceVersion < 2)
 	{
 		NSLog(@"Updating services menu…");
@@ -308,7 +308,7 @@
 
                 else {
 					
-                    int alertReturn = NSRunAlertPanel(nil, @"Could not save changes while quitting. Quit anyway?" , @"Quit anyway", @"Cancel", nil);
+                    NSInteger alertReturn = NSRunAlertPanel(nil, @"Could not save changes while quitting. Quit anyway?" , @"Quit anyway", @"Cancel", nil);
                     if (alertReturn == NSAlertAlternateReturn) {
                         reply = NSTerminateCancel;	
                     }
